update calling spec, allow pointer returns, make show work

+++ internal/types/defaults.go
@@ -12,7 +12,7 @@ var Boolean = &boolean{}
 // Pict is shorthand for [,]{float, float, float, float}
 var Pict = &Array{
 	Inner: &Tuple{
-		Types: []Type{Integer, Integer, Integer, Float},
+		Types: []Type{Float, Float, Float, Float},
 	},
 	Rank: 2,
 }
